Fix: Add a debounce function to the user input
By choice I query for 5 row per user input, the api call is fired with
every input change wich resulted in 5 results for every single input.
The solution of emptyListContainer() worked at a certain degree and
failed with rapid inputs.
This fix brings a debouncing function to mitigate this bug.